Add files via upload
This is a simple python code for a simple pong game. To test this game install python to your computer and run the code on the terminal.
Left player uses "w" key to move the paddle up and "s" key to move the paddle down
right player uses up-arrow to move the paddle up and down-arrow to move the paddle down
use "q" key to quit the game and "r" key to reset the game while playing
the game is won if one of the two sides reach 21 points and maintain a 2 point lead


I made this code using perplexity AI, an LLM. I've never developed a game before and never coded past a calculator. 
I made this game for AI week in TKS and there are few things I would like to share regarding AI and coding. 
AI is a great tool to use when coding so knowing how to use it is a real game changer. It all comes down to your prompting skills!
To use AI effectively and induce correct responses, we need to be good at prompting( basically communicating with the AI)!
We can use AI to significantly lower coding time and also debugging time.